Changes to enable libxipc compilation on gcc 3.2.1 and gcc 2.9x.
git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@121 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/libxipc/xrl_pf_stcp.hh
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-// $XORP: xorp/libxipc/xrl_pf_stcp.hh,v 1.2 2002/12/18 22:54:30 hodson Exp $
+// $XORP: xorp/libxipc/xrl_pf_stcp.hh,v 1.3 2002/12/19 01:29:14 hodson Exp $
 
 #ifndef __XRLPF_STCP_HH__
 #define __XRLPF_STCP_HH__
@@ -46,7 +46,7 @@ struct RequestState {
 
 class XrlPFSTCPListener : public XrlPFListener {
 public:
-    XrlPFSTCPListener(EventLoop& e, XrlCmdMap* m = 0, int port = 0)
+    XrlPFSTCPListener(EventLoop& e, XrlCmdMap* m = 0, uint16_t port = 0)
 	throw (XrlPFConstructorError);
     ~XrlPFSTCPListener();
